fix check-format
