+++ app/(inferenceai)/inferenceai/components/partner-list.tsx
@@ -6,13 +6,13 @@ export default function PartnersList() {
   const { header, logos } = data.partners;
 
   return (
-    <section className="relative w-full py-16 overflow-hidden bg-[#050013]">
-      <div className="relative px-4 md:px-20 z-10 mx-auto flex flex-col md:flex-row md:items-center justify-center md:justify-between gap-y-8">
-        <h2 className="text-white text-2xl xl:text-[32px] font-medium">
+    <section className="relative w-full overflow-hidden bg-[#050013] py-16">
+      <div className="relative z-10 mx-auto flex flex-col justify-center gap-y-8 px-4 md:flex-row md:items-center md:justify-between md:px-20">
+        <h2 className="text-2xl font-medium text-white xl:text-[32px]">
           {header}
         </h2>
 
-        <div className="grid grid-cols-3 gap-4 md:gap-6 items-center justify-center md:justify-end">
+        <div className="grid grid-cols-3 items-center justify-center gap-4 md:justify-end md:gap-6">
           {logos.map((partner, index) => (
             <div key={index} className="flex w-auto items-center justify-start">
               <Image
@@ -21,8 +21,8 @@ export default function PartnersList() {
                 height={36}
                 width={120}
                 className={cn(
-                  "object-contain h-[30px] w-auto",
-                  "grayscale opacity-60 hover:opacity-100 transition"
+                  "h-[30px] w-auto object-contain",
+                  "opacity-60 grayscale transition hover:opacity-100"
                 )}
               />
             </div>
