+++ app/(inferenceai)/inferenceai/layout.tsx
@@ -4,7 +4,7 @@ import Hero from "./components/hero";
 import Footer from "./components/footer";
 
 export default function MainInferenceAILayout({
-  children,
+  children
 }: {
   children: React.ReactNode;
 }) {
