+++ app/globals.css
@@ -79,7 +79,8 @@ body {
 
 /* inferenceai AI */
 .bg-grid-faqs {
-  background: linear-gradient(0deg, #050013, #050013),
+  background:
+    linear-gradient(0deg, #050013, #050013),
     linear-gradient(180deg, #1513374d 0%, #15133700 23.58%);
 }
 
@@ -122,7 +123,9 @@ body {
     #877f9f 0%,
     rgba(135, 127, 159, 0.4) 100%
   );
-  -webkit-mask: linear-gradient(#fff 0 0) content-box, linear-gradient(#fff 0 0);
+  -webkit-mask:
+    linear-gradient(#fff 0 0) content-box,
+    linear-gradient(#fff 0 0);
   -webkit-mask-composite: xor;
   mask-composite: exclude;
   pointer-events: none;
